+++ example/celery+django/celery_django/celery.py
@@ -23,6 +23,7 @@ bugsnag.configure(**bugsnagconfig)
 
 connect_failure_handler()
 
+
 @app.task(bind=True)
 def debug_task(self):
     print('Request: {0!r}'.format(self.request))