+++ example/django21/bugsnag_demo/urls.py
@@ -1,6 +1,12 @@
 from django.conf.urls import url
-from demo.views import (index, crash, crash_with_callback, handled,
-                        context, notify_meta)
+from demo.views import (
+    index,
+    crash,
+    crash_with_callback,
+    handled,
+    context,
+    notify_meta,
+)
 
 urlpatterns = [
     url(r'^$', index),